requirements.txt
coverage==4.2
hypothesis==3.6.0
pycrypto==2.6.1
pytest==3.0.4
pytest-cov==2.4.0